# docs/api/openapi.yaml
openapi: 3.1.1
info:
  title: Mobily Transformation Hub API
  version: 1.0.0-p1
  summary: P1 contract (identity, scoped access, organizations, transformations, audit).
  description: |
    Versioned REST contract for the Mobily Transformation Hub (T-DG1-ARCH-01, ADR-0007).

    **Source of truth.** This file is the contract. It is owned by solution-architect and changed only
    through the orchestrator. The zod schemas in `@mth/shared/schemas` mirror it, and API contract tests
    validate every response of the route under test against this document (ADR-0007).

    **Versioning.** All business endpoints live under `/api/v1`. Additive changes (new optional fields,
    new endpoints, new enum values in response-only fields) are allowed within v1. Breaking changes need
    `/api/v2` or a reopened gate (agent-protocol: contracts stay stable once a gate approves them).

    **Authentication (ADR-0005).** Browser clients authenticate with OpenID Connect authorization code +
    PKCE against the corporate identity provider (Keycloak is the portable test IdP). The API is the
    confidential OIDC client; tokens never reach the browser. After the callback the API creates a
    server-side session in PostgreSQL and sets an HttpOnly, Secure, SameSite=Lax cookie. In production
    (HTTPS) the cookie is `__Host-mth_session`; with an `http://localhost` base URL in development it is
    `mth_session`. A local **dev login** exists only when `AUTH_MODE=dev`. The server refuses to start
...

...

...

...
servers:
...
security:
  - sessionCookie: []
paths:
...
  /api/v1/auth/login:
    get:
      tags: [auth]
      operationId: startOidcLogin
      summary: Start OIDC authorization code + PKCE login.
      description: |
        Creates a short-lived server-side login state (state, nonce, PKCE code_verifier, returnTo) and
        redirects to the identity provider's authorization endpoint. Answers 404 when OIDC is not configured.
      security: []
      parameters:
        - name: returnTo
          in: query
          required: false
          description: Same-origin relative path to return to after login. Absolute or protocol-relative URLs are rejected (open-redirect protection).
          schema: { type: string, pattern: "^/(?!/)[^\\s]*$", maxLength: 2048 }
      responses:
        "302":
          description: Redirect to the identity provider.
          headers:
            Location: { $ref: "#/components/headers/Location" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/auth/callback:
    get:
      tags: [auth]
      operationId: completeOidcLogin
      summary: OIDC redirect URI. Exchanges the code, binds the identity, creates the session.
      description: |
        Validates `state` against the stored login state (single use, expires after 10 minutes), exchanges
        the code with the PKCE verifier, validates the ID token (issuer, audience, nonce, expiry), and
        resolves the user by the (issuer, subject) pair (ADR-0005). It then creates the session, sets the session
        cookie and redirects to `returnTo`. Failures redirect to `/login?error=<code>` and never render IdP
        error text verbatim.
      security: []
      parameters:
        - { name: code, in: query, required: false, schema: { type: string, maxLength: 4096 } }
        - { name: state, in: query, required: true, schema: { type: string, maxLength: 512 } }
        - { name: error, in: query, required: false, schema: { type: string, maxLength: 256 } }
      responses:
        "302":
          description: Redirect to the application (success) or to `/login?error=<code>` (failure).
          headers:
            Location: { $ref: "#/components/headers/Location" }
            Set-Cookie: { $ref: "#/components/headers/Set-Cookie" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/auth/dev-login:
    post:
      tags: [auth]
      operationId: devLogin
      summary: Development-only local login (AUTH_MODE=dev).
      description: |
        Available only when `AUTH_MODE=dev`, and never with `NODE_ENV=production` (startup refuses that
        combination). In every other mode it answers 404 as if the route did not exist. It signs in an
        existing **synthetic** development user bound to issuer `urn:mth:dev-local`. It never creates users
        and never grants roles. Needs a same-origin `Origin` header (login-CSRF protection).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DevLoginRequest" }
      responses:
        "204":
          description: Signed in; session cookie set.
          headers:
            Set-Cookie: { $ref: "#/components/headers/Set-Cookie" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: End the server-side session.
      description: Revokes the session row, clears the cookie and returns the IdP end-session URL when OIDC is in use.
      security:
        - sessionCookie: []
          csrfToken: []
      responses:
        "200":
          description: Signed out.
          headers:
            Set-Cookie: { $ref: "#/components/headers/Set-Cookie" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LogoutResult" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/me:
    get:
      tags: [me]
      operationId: getMe
      summary: Current user, CSRF token, assignments and effective permissions per scope.
      responses:
        "200":
          description: The signed-in principal.
          headers:
            X-Request-Id: { $ref: "#/components/headers/X-Request-Id" }
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Me" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
  /api/v1/me/preferences:
    put:
      tags: [me]
      operationId: updateMyPreferences
      summary: Persist the language and timezone preference (persistent language switch, REQ-S15-007).
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PreferencesUpdate" }
      responses:
        "200":
          description: Updated user.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/VersionConflict" }
        "428": { $ref: "#/components/responses/PreconditionRequired" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-mth_session
      description: |
        Opaque server-side session ID (256-bit random; only its SHA-256 is stored). HttpOnly, Secure,
        SameSite=Lax, Path=/. Named `mth_session` (without the `__Host-` prefix) only for an http://localhost base URL in development.
        Idle timeout SESSION_IDLE_MINUTES (default 30), absolute SESSION_ABSOLUTE_HOURS (default 10).
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Per-session synchronizer token from `GET /api/v1/me` (`csrfToken`). Required on every unsafe method, together with a same-origin `Origin` header.
  parameters:
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: The ETag (`"<version>"`) of the representation being changed. A missing header gives 428; a stale one gives 409.
      schema: { type: string, pattern: '^"[1-9][0-9]{0,9}"$' }
...
  headers:
    X-Request-Id:
      description: Request correlation ID (echoed from a valid inbound header or generated), also in every log line and audit event.
      schema: { type: string, maxLength: 128 }
    ETag:
      description: Strong entity tag `"<version>"`.
      schema: { type: string, pattern: '^"[1-9][0-9]{0,9}"$' }
    Location:
      schema: { type: string }
    Set-Cookie:
      schema: { type: string }
    Retry-After:
      schema: { type: integer, minimum: 0 }
  responses:
    ValidationError:
      description: Request failed schema validation (`urn:mth:problem:validation`), with field errors.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unauthenticated:
      description: No valid session (`urn:mth:problem:unauthenticated`).
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Forbidden:
      description: The caller can see the resource but lacks the permission, or the CSRF/Origin check failed (`forbidden` or `csrf`).
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: The resource does not exist or the caller may not read it (`urn:mth:problem:not-found`).
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    VersionConflict:
      description: If-Match is stale (`urn:mth:problem:version-conflict`). `currentVersion` holds the server's version; nothing was written.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    PreconditionRequired:
      description: If-Match header missing (`urn:mth:problem:precondition-required`).
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    RateLimited:
      description: Too many requests (`urn:mth:problem:rate-limited`).
      headers:
        Retry-After: { $ref: "#/components/headers/Retry-After" }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Uuid:
      type: string
      format: uuid
      description: UUIDv7 (time-ordered), generated by the application.
...
    Locale:
      type: string
      enum: [ar, en]
    TimeZone:
      type: string
      description: IANA time zone name, validated against the runtime's zone database.
      minLength: 1
      maxLength: 64
      examples: [Asia/Riyadh]
...
    Name:
      type: string
      minLength: 1
      maxLength: 200
...
    UserStatus:
      type: string
      enum: [active, disabled]
...
    ScopeRef:
      type: object
      required: [type, id]
      additionalProperties: false
      properties:
        type: { $ref: "#/components/schemas/ScopeType" }
        id: { $ref: "#/components/schemas/Uuid" }
    PermissionCode:
      type: string
      enum:
        - organization.read
        - organization.manage
        - business_unit.read
        - business_unit.manage
        - user.read
        - user.manage
        - role.read
        - access.read
        - access.assign
        - transformation.read
        - transformation.create
        - transformation.update
        - transformation.archive
        - audit.read
        - gate.decide
        - finance.validate
...
    Version:
      type: integer
      minimum: 1
      description: Optimistic-concurrency version; starts at 1 and increases by exactly 1 per change.
    Problem:
      type: object
      description: RFC 9457 problem details.
      required: [type, title, status, code, requestId]
      properties:
        type:
          type: string
          enum:
            - urn:mth:problem:validation
            - urn:mth:problem:unauthenticated
            - urn:mth:problem:forbidden
            - urn:mth:problem:not-found
            - urn:mth:problem:version-conflict
            - urn:mth:problem:precondition-required
            - urn:mth:problem:invalid-transition
            - urn:mth:problem:duplicate
            - urn:mth:problem:csrf
            - urn:mth:problem:rate-limited
            - urn:mth:problem:unavailable
            - urn:mth:problem:internal
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string }
        code:
          type: string
          description: Stable i18n key suffix (`problem.<code>`), e.g. `version_conflict`.
          pattern: "^[a-z][a-z0-9_.]*$"
        requestId: { type: string }
        errors:
          type: array
          items: { $ref: "#/components/schemas/FieldError" }
        currentVersion: { $ref: "#/components/schemas/Version" }
...
    DevLoginRequest:
      type: object
      required: [username]
      additionalProperties: false
      properties:
        username: { type: string, pattern: "^[a-z0-9._-]{2,64}$", description: "Subject of a seeded synthetic dev user." }
    LogoutResult:
      type: object
      required: [endSessionUrl]
      additionalProperties: false
      properties:
        endSessionUrl:
          type: [string, "null"]
          description: IdP RP-initiated logout URL to navigate to, or null (dev mode / no end_session_endpoint).
    Me:
      type: object
      required: [user, authMode, csrfToken, productName, organization, assignments, effectivePermissions]
      additionalProperties: false
      properties:
        user: { $ref: "#/components/schemas/User" }
        authMode: { type: string, enum: [oidc, dev] }
        csrfToken: { type: string, minLength: 32 }
        productName: { type: string, description: "Configurable product name (REQ-S01-002)." }
        organization:
          description: The user's home organization and its defaults.
          $ref: "#/components/schemas/Organization"
        assignments:
          type: array
          description: Active (non-revoked, currently effective) assignments of the user.
          items: { $ref: "#/components/schemas/RoleAssignment" }
        effectivePermissions:
          type: array
          description: |
            Permissions per granted scope, computed by the same policy function the server uses. A UI hint only;
            the server re-checks every request. When a role inherits downward, `inheritsDownward` is true and the
            permissions also apply to every scope inside `scope`.
          items:
            type: object
            required: [scope, inheritsDownward, permissions]
            additionalProperties: false
            properties:
              scope: { $ref: "#/components/schemas/ScopeRef" }
              inheritsDownward: { type: boolean }
              permissions: { type: array, uniqueItems: true, items: { $ref: "#/components/schemas/PermissionCode" } }
    PreferencesUpdate:
      type: object
      minProperties: 1
      additionalProperties: false
      properties:
        preferredLocale: { $ref: "#/components/schemas/Locale" }
        timezone:
          oneOf:
            - { $ref: "#/components/schemas/TimeZone" }
            - { type: "null" }
    Organization:
      type: object
      required: [id, code, nameEn, nameAr, defaultTimezone, defaultCurrency, defaultLocale, status, version, createdAt, updatedAt]
      additionalProperties: false
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        code: { $ref: "#/components/schemas/Code" }
        nameEn: { $ref: "#/components/schemas/Name" }
        nameAr: { $ref: "#/components/schemas/Name" }
        defaultTimezone: { $ref: "#/components/schemas/TimeZone" }
        defaultCurrency: { $ref: "#/components/schemas/Currency" }
        defaultLocale: { $ref: "#/components/schemas/Locale" }
        status: { $ref: "#/components/schemas/ActiveStatus" }
        version: { $ref: "#/components/schemas/Version" }
        createdAt: { $ref: "#/components/schemas/Timestamp" }
        updatedAt: { $ref: "#/components/schemas/Timestamp" }
...
    UserIdentity:
      type: object
      required: [issuer, subject, createdAt, lastLoginAt]
      additionalProperties: false
      properties:
        issuer: { type: string, maxLength: 512 }
        subject: { type: string, maxLength: 255 }
        createdAt: { $ref: "#/components/schemas/Timestamp" }
        lastLoginAt:
          oneOf:
            - { $ref: "#/components/schemas/Timestamp" }
            - { type: "null" }
    User:
      type: object
      required: [id, organizationId, displayName, email, preferredLocale, timezone, status, identities, version, createdAt, updatedAt]
      additionalProperties: false
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        organizationId: { $ref: "#/components/schemas/Uuid" }
        displayName: { $ref: "#/components/schemas/Name" }
        email: { type: [string, "null"], format: email, maxLength: 320 }
        preferredLocale: { $ref: "#/components/schemas/Locale" }
        timezone:
          description: Personal override; null means the organization default.
          oneOf:
            - { $ref: "#/components/schemas/TimeZone" }
            - { type: "null" }
        status: { $ref: "#/components/schemas/UserStatus" }
        identities:
          type: array
          description: Bound IdP identities (issuer + subject). Shown to user.read holders and to the user themself.
          items: { $ref: "#/components/schemas/UserIdentity" }
        version: { $ref: "#/components/schemas/Version" }
        createdAt: { $ref: "#/components/schemas/Timestamp" }
        updatedAt: { $ref: "#/components/schemas/Timestamp" }
...
    RoleAssignment:
      type: object
      required:
        [id, organizationId, userId, roleCode, scope, effectiveFrom, effectiveTo, reason, grantedBy, revokedAt, revokedBy, revokeReason, version, createdAt]
      additionalProperties: false
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        organizationId: { $ref: "#/components/schemas/Uuid" }
        userId: { $ref: "#/components/schemas/Uuid" }
        roleCode: { $ref: "#/components/schemas/RoleCode" }
        scope: { $ref: "#/components/schemas/ScopeRef" }
        effectiveFrom: { $ref: "#/components/schemas/Timestamp" }
        effectiveTo:
          oneOf:
            - { $ref: "#/components/schemas/Timestamp" }
            - { type: "null" }
        reason: { $ref: "#/components/schemas/Reason" }
        grantedBy: { $ref: "#/components/schemas/Uuid" }
        revokedAt:
          oneOf:
            - { $ref: "#/components/schemas/Timestamp" }
            - { type: "null" }
        revokedBy:
          oneOf:
            - { $ref: "#/components/schemas/Uuid" }
            - { type: "null" }
        revokeReason: { type: [string, "null"] }
        version: { $ref: "#/components/schemas/Version" }
        createdAt: { $ref: "#/components/schemas/Timestamp" }
...